+++ src/main/java/it/unibo/pensilina14/bullet/ballet/characters/Player.java
@@ -32,7 +32,7 @@ public class Player implements Characters{
     }
 
     @Override
-    public void setHealth(float setHealth) {
+    public void setHealth(double setHealth) {
         this.health = setHealth;
     }
